Split ToolsList component into sub-components for better readability
